+++ mobile/lib/controller/auth_controller.dart
@@ -3,10 +3,10 @@ import 'package:flutter/material.dart';
 import 'package:http/http.dart' as http;
 import 'package:image_picker/image_picker.dart';
 import 'package:mobile/controller/networking_controller.dart';
-import 'package:mobile/view/auth_pages/components/custom_dialogue.dart';
-import 'package:mobile/view/auth_pages/pages/auth/profile_info_photo.dart';
-import 'package:mobile/view/auth_pages/pages/auth/welcome_page.dart';
-import 'package:mobile/view/auth_pages/pages/home/home_page.dart';
+import 'package:mobile/view/auth/components/custom_dialogue.dart';
+import 'package:mobile/view/auth/pages/profile_info_photo.dart';
+import 'package:mobile/view/auth/pages/welcome_page.dart';
+import 'package:mobile/view/home/pages/home_page.dart';
 import 'dart:convert';
 import 'local_data_controller.dart';
 import 'dart:io';
