+++ mobile/lib/view/auth/components/custom_positioned_text.dart
