+++ mobile/lib/view/auth/components/search_bar_widget.dart
