+++ mobile/lib/view/auth/pages/welcome_page.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
-import 'package:mobile/view/auth_pages/pages/auth/login_page.dart';
-import 'package:mobile/view/auth_pages/pages/auth/register_page.dart';
-import '../../components/action_button.dart';
+import 'package:mobile/view/auth/pages/login_page.dart';
+import 'package:mobile/view/auth/pages/register_page.dart';
+import '../components/action_button.dart';
 import 'package:mobile/model/paths.dart';
 import 'package:mobile/model/app_text_data.dart';
 import 'package:mobile/controller/auth_controller.dart';
