+++ mobile/lib/view/home/components/featured_card.dart
