+++ mobile/lib/view/home/components/workshop_cell.dart
@@ -2,7 +2,7 @@ import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
 import 'package:mobile/model/workshop_model.dart';
-import 'package:mobile/view/auth_pages/pages/home/home_page.dart';
+import 'package:mobile/view/home/pages/home_page.dart';
 
 class WorkshopCell extends StatelessWidget {
   const WorkshopCell({super.key, required this.workshop, required this.index});
